third_party/WebKit/LayoutTests/http/tests/devtools/sass/test-mapping-good-expected.txt

Verify WI.ASTSourceMap rebaselines.

Mapped CSS: 16
Mapped SCSS: 16
No stale entries found.

Running: testRebase
Mapped CSS: 16
Mapped SCSS: 16
No stale entries found.
